+++ R/enricher.R
@@ -12,7 +12,7 @@
 ##' @param TERM2GENE user input annotation of TERM TO GENE mapping, a data.frame of 2 column with term and gene
 ##' @param TERM2NAME user input of TERM TO NAME mapping, a data.frame of 2 column with term and name
 ##' @return A \code{enrichResult} instance
-##' @author Guangchuang Yu
+##' @author Guangchuang Yu \url{https://yulab-smu.top}
 ##' @export
 enricher <- function(gene,
                      pvalueCutoff = 0.05,
@@ -53,7 +53,7 @@ enricher <- function(gene,
 ##' @param by one of 'fgsea' or 'DOSE'
 ##' @param ... other parameter
 ##' @return gseaResult object
-##' @author Guangchuang Yu
+##' @author Guangchuang Yu \url{https://yulab-smu.top}
 ##' @export
 GSEA <- function(geneList,
                  exponent = 1,